Added habit form to navbar and fixed habit form bug
For testing and efficiency purposes, I added the habit form to the navbar. Once the habit page is developed, this form will be removed from the navbar. I also fixed a bug in the habit form where the period would not be filled in with daily if the user does not manually change it, so I made the form autofill daily if the user does not change it.