Implemented the generate buddy link feature
In the header, I added the link to the generate buddy link page. I also create the component for the page and connected to the backend to fill the input field with the generate link. Also, i installed the copy-to-clipboard dependecy from react so the user can copy and paste their links